huge refactoring, declare if not declared, change style, type improvments

+++ test/optimizer-test.cpp
@@ -4,9 +4,9 @@
 
 #include <catch2/catch.hpp>
 
-#include <nav2_costmap_2d/cost_values.hpp>
-#include "nav2_costmap_2d/costmap_2d_ros.hpp"
 #include "nav2_costmap_2d/costmap_2d.hpp"
+#include "nav2_costmap_2d/costmap_2d_ros.hpp"
+#include <nav2_costmap_2d/cost_values.hpp>
 
 #include <rclcpp/executors.hpp>
 
@@ -17,25 +17,22 @@
 #include "mppi/Models.hpp"
 #include "mppi/impl/Optimizer.hpp"
 
-#include "xtensor/xarray.hpp"
-#include <xtensor/xview.hpp>
-#include <xtensor/xio.hpp>
-#include <iostream>
 #include "tests_utils.hpp"
-
-
+#include "xtensor/xarray.hpp"
+#include <iostream>
+#include <xtensor/xio.hpp>
+#include <xtensor/xview.hpp>
 
 /**
  * Adds some parameters for the optimizer to a special container.
  *
  * @param params_ container for optimizer's parameters.
-*/
-void setUpOptimizerParams(std::vector<rclcpp::Parameter> &params_){
+ */
+void setUpOptimizerParams(std::vector<rclcpp::Parameter> &params_) {
   params_.push_back(rclcpp::Parameter("TestNode.iteration_count", 5));
   params_.push_back(rclcpp::Parameter("TestNode.lookahead_dist", 5));
   params_.push_back(rclcpp::Parameter("TestNode.time_steps", 30));
 }
-
 
 TEST_CASE("Optimizer evaluates Next Control", "") {
   using T = float;
@@ -43,19 +40,19 @@ TEST_CASE("Optimizer evaluates Next Control", "") {
   std::string node_name = "TestNode";
   auto node = std::make_shared<rclcpp_lifecycle::LifecycleNode>(node_name);
 
-  auto costmap_ros = std::make_shared<nav2_costmap_2d::Costmap2DROS>("cost_map_node");
+  auto costmap_ros =
+      std::make_shared<nav2_costmap_2d::Costmap2DROS>("cost_map_node");
   auto st = rclcpp_lifecycle::State{};
 
-  auto & model = mppi::models::NaiveModel<T>;
+  auto &model = mppi::models::NaiveModel<T>;
 
-  auto optimizer =
-    mppi::optimization::Optimizer<T>();
+  auto optimizer = mppi::optimization::Optimizer<T>();
 
   costmap_ros->on_configure(st);
   optimizer.on_configure(node, node_name, costmap_ros, model);
   optimizer.on_activate();
 
-  size_t poses_count = GENERATE(10, 30, 100);
+  unsigned int poses_count = GENERATE(10U, 30U, 100U);
 
   SECTION("Running evalNextControl") {
     geometry_msgs::msg::Twist twist;
@@ -63,10 +60,10 @@ TEST_CASE("Optimizer evaluates Next Control", "") {
     std::string frame = "odom";
     auto time = node->get_clock()->now();
 
-    auto setHeader = [&](auto && msg) {
-        msg.header.frame_id = frame;
-        msg.header.stamp = time;
-      };
+    auto setHeader = [&](auto &&msg) {
+      msg.header.frame_id = frame;
+      msg.header.stamp = time;
+    };
 
     nav_msgs::msg::Path path;
     geometry_msgs::msg::PoseStamped ps;
@@ -74,10 +71,10 @@ TEST_CASE("Optimizer evaluates Next Control", "") {
     setHeader(path);
 
     auto fillPath = [&](size_t count) {
-        for (size_t i = 0; i < count; i++) {
-          path.poses.push_back(ps);
-        }
-      };
+      for (size_t i = 0; i < count; i++) {
+        path.poses.push_back(ps);
+      }
+    };
 
     fillPath(poses_count);
 
@@ -97,7 +94,6 @@ TEST_CASE("Optimizer evaluates Next Control", "") {
   costmap_ros.reset();
 }
 
-
 TEST_CASE("Optimizer with costmap2d and obstacles", "[collision]") {
 
   using T = float;
@@ -116,55 +112,52 @@ TEST_CASE("Optimizer with costmap2d and obstacles", "[collision]") {
   double origin_x = 0.0;
   double origin_y = 0.0;
   unsigned char default_value = 0;
-  
+
   // args for obstacle on costmap2d
-  const unsigned int upper_left_corner_x = 5;  
-  const unsigned int upper_left_corner_y = 7;  
-  const unsigned int obstacle_side_size_cells = 9;  
-  unsigned char obstacle_cost = 255;                                     
-  
+  const unsigned int upper_left_corner_x = 5;
+  const unsigned int upper_left_corner_y = 7;
+  const unsigned int obstacle_side_size_cells = 9;
+  unsigned char obstacle_cost = 255;
+
   // create parameters for reference path generation
   size_t poses_count = 40;
-  float x_step = 0.015;
-  float y_step = 0.035;
+  float x_step = 0.015f;
+  float y_step = 0.035f;
 
-  auto node = std::make_shared<rclcpp_lifecycle::LifecycleNode>(node_name, options);
-  auto costmap_ros = std::make_shared<nav2_costmap_2d::Costmap2DROS>("cost_map_node");
-  
-  auto & model = mppi::models::NaiveModel<T>;
+  auto node =
+      std::make_shared<rclcpp_lifecycle::LifecycleNode>(node_name, options);
+  auto costmap_ros =
+      std::make_shared<nav2_costmap_2d::Costmap2DROS>("cost_map_node");
+
+  auto &model = mppi::models::NaiveModel<T>;
   auto optimizer = mppi::optimization::Optimizer<T>();
 
   costmap_ros->on_configure(state);
   *costmap_ros->getCostmap() = *std::make_shared<nav2_costmap_2d::Costmap2D>(
-    cells_size_x, 
-    cells_size_y,
-    resolution,
-    origin_x,
-    origin_y,
-    default_value
-  );
+      cells_size_x, cells_size_y, resolution, origin_x, origin_y,
+      default_value);
 
   // creating a square obstacle
-  addObstacle(*costmap_ros->getCostmap(), upper_left_corner_x, 
+  addObstacle(*costmap_ros->getCostmap(), upper_left_corner_x,
               upper_left_corner_y, obstacle_side_size_cells, obstacle_cost);
 
   optimizer.on_configure(node, node_name, costmap_ros, model);
   optimizer.on_activate();
 
-  float start_point_x = GENERATE(1.0, 0.4);
-  float start_point_y = 0.4;
+  float start_point_x = GENERATE(1.0f, 0.4f);
+  float start_point_y = 0.4f;
 
   SECTION("evalNextBestControl doesn't produce path crossing the obstacles") {
-    
-    std::string frame = "odom";              // frame for header in path and points
-    auto time = node->get_clock()->now();    // time for header in path
+
+    std::string frame = "odom";           // frame for header in path and points
+    auto time = node->get_clock()->now(); // time for header in path
 
     nav_msgs::msg::Path reference_path;
     geometry_msgs::msg::PoseStamped reference_goal_pose;
-    geometry_msgs::msg::PoseStamped init_robot_pose;   
+    geometry_msgs::msg::PoseStamped init_robot_pose;
     init_robot_pose.pose.position.x = start_point_x;
-    init_robot_pose.pose.position.y = start_point_y;          
-    geometry_msgs::msg::Twist init_robot_vel;           
+    init_robot_pose.pose.position.y = start_point_y;
+    geometry_msgs::msg::Twist init_robot_vel;
 
     // lambda expression for setting header
     auto setHeader = [&](auto &&msg) {
@@ -175,11 +168,13 @@ TEST_CASE("Optimizer with costmap2d and obstacles", "[collision]") {
     // lambda expression for refernce path generation
     auto fillRealPath = [&](size_t count) {
       for (size_t i = 0; i < count; i++) {
-          reference_goal_pose.pose.position.x = i*x_step + start_point_x;
-          reference_goal_pose.pose.position.y = i*y_step + start_point_y;
-          reference_path.poses.push_back(reference_goal_pose);
+        reference_goal_pose.pose.position.x =
+            static_cast<float>(i) * x_step + start_point_x;
+        reference_goal_pose.pose.position.y =
+            static_cast<float>(i) * y_step + start_point_y;
+        reference_path.poses.push_back(reference_goal_pose);
       }
-    };  
+    };
 
     setHeader(reference_goal_pose);
     setHeader(init_robot_pose);
@@ -188,12 +183,15 @@ TEST_CASE("Optimizer with costmap2d and obstacles", "[collision]") {
     fillRealPath(poses_count);
 
     // update controal sequence in optimizer
-    CHECK_NOTHROW(optimizer.evalNextBestControl(init_robot_pose, init_robot_vel, reference_path));
+    CHECK_NOTHROW(optimizer.evalNextBestControl(init_robot_pose, init_robot_vel,
+                                                reference_path));
     // get best trajectory from optimizer
-    auto trajectory = optimizer.evalTrajectoryFromControlSequence(init_robot_pose, init_robot_vel);
+    auto trajectory = optimizer.evalTrajectoryFromControlSequence(
+        init_robot_pose, init_robot_vel);
 #ifdef TEST_DEBUG_INFO
-    printMapWithGoalAndTrajectory(*costmap_ros->getCostmap(), trajectory, reference_goal_pose);
-#endif   
+    printMapWithGoalAndTrajectory(*costmap_ros->getCostmap(), trajectory,
+                                  reference_goal_pose);
+#endif
     CHECK(!inCollision(trajectory, *costmap_ros->getCostmap()));
   }
 
@@ -201,5 +199,4 @@ TEST_CASE("Optimizer with costmap2d and obstacles", "[collision]") {
   optimizer.on_cleanup();
   costmap_ros->on_cleanup(state);
   costmap_ros.reset();
-
 }
